Drop support for Python 3.4
3.4 is over 4 years old, and Python 3.7 has been released.
Let's move forward.
