+++ README.rst
@@ -46,7 +46,7 @@ Install
 
     pip install -U webargs
 
-webargs supports Python >= 2.7 or >= 3.4.
+webargs supports Python >= 2.7 or >= 3.5.
 
 
 Documentation
