+++ setup.cfg
@@ -4,4 +4,4 @@ universal = 1
 [flake8]
 ignore = E127,E128,E265,E302,N803,N804,N806,E731,E402,E266,E305
 max-line-length = 100
-exclude = .git,.ropeproject,.tox,docs,.git,setup.py,tests/compat.py,build,.direnv
+exclude = .git,.ropeproject,.tox,build,.direnv,__pycache__
