+++ tests/compat.py
@@ -1,13 +1,14 @@
 # -*- coding: utf-8 -*-
+# flake8: noqa
 import sys
 
 PY2 = int(sys.version[0]) == 2
 
 if PY2:
-    text_type = unicode  # flake8: noqa
+    text_type = unicode
     binary_type = str
-    string_types = (str, unicode)  # flake8: noqa
-    basestring = basestring  # flake8: noqa
+    string_types = (str, unicode)
+    basestring = basestring
 else:
     text_type = str
     binary_type = bytes
