+++ webargs/asyncparser.py
@@ -1,21 +1,15 @@
 # -*- coding: utf-8 -*-
-"""Asynchronous request parser. Compatible with Python>=3.4."""
+"""Asynchronous request parser. Compatible with Python>=3.5."""
 import asyncio
-import sys
+import collections
+import functools
 import inspect
 
 import marshmallow as ma
-from marshmallow.compat import iteritems
 from marshmallow.utils import missing
 
 from webargs import core
 
-PY_34 = sys.version_info < (3, 5)
-
-if PY_34:
-    from webargs.async_decorators34 import _use_args
-else:
-    from webargs.async_decorators import _use_args
 
 class AsyncParser(core.Parser):
     """Asynchronous variant of `webargs.core.Parser`, where parsing methods may be
@@ -40,7 +34,7 @@ class AsyncParser(core.Parser):
         else:
             argdict = schema.fields
             parsed = {}
-            for argname, field_obj in iteritems(argdict):
+            for argname, field_obj in argdict.items():
                 if core.MARSHMALLOW_VERSION_INFO[0] < 3:
                     parsed_value = yield from self.parse_arg(argname, field_obj, req, locations)
                     # If load_from is specified on the field, try to parse from that key
@@ -80,7 +74,66 @@ class AsyncParser(core.Parser):
             core.fill_in_missing_args(data, schema)
         return data
 
-    use_args = _use_args
+    def use_args(self, argmap, req=None, locations=None, as_kwargs=False, validate=None):
+        """Decorator that injects parsed arguments into a view function or method.
+
+        Receives the same arguments as `webargs.core.Parser.use_args`.
+        """
+        locations = locations or self.locations
+        request_obj = req
+        # Optimization: If argmap is passed as a dictionary, we only need
+        # to generate a Schema once
+        if isinstance(argmap, collections.Mapping):
+            argmap = core.argmap2schema(argmap)()
+
+        def decorator(func):
+            req_ = request_obj
+
+            if inspect.iscoroutinefunction(func):
+                @functools.wraps(func)
+                async def wrapper(*args, **kwargs):
+                    req_obj = req_
+
+                    # if as_kwargs is passed, must include all args
+                    force_all = as_kwargs
+
+                    if not req_obj:
+                        req_obj = self.get_request_from_view_args(func, args, kwargs)
+                    # NOTE: At this point, argmap may be a Schema, callable, or dict
+                    parsed_args = await self.parse(argmap,
+                                                        req=req_obj, locations=locations,
+                                                        validate=validate, force_all=force_all)
+                    if as_kwargs:
+                        kwargs.update(parsed_args)
+                        return await func(*args, **kwargs)
+                    else:
+                        # Add parsed_args after other positional arguments
+                        new_args = args + (parsed_args, )
+                        return await func(*new_args, **kwargs)
+            else:
+                @functools.wraps(func)
+                def wrapper(*args, **kwargs):
+                    req_obj = req_
+
+                    # if as_kwargs is passed, must include all args
+                    force_all = as_kwargs
+
+                    if not req_obj:
+                        req_obj = self.get_request_from_view_args(func, args, kwargs)
+                    # NOTE: At this point, argmap may be a Schema, callable, or dict
+                    parsed_args = yield from self.parse(argmap,
+                                                        req=req_obj, locations=locations,
+                                                        validate=validate, force_all=force_all)
+                    if as_kwargs:
+                        kwargs.update(parsed_args)
+                        return func(*args, **kwargs)
+                    else:
+                        # Add parsed_args after other positional arguments
+                        new_args = args + (parsed_args, )
+                        return func(*new_args, **kwargs)
+            wrapper.__wrapped__ = func
+            return wrapper
+        return decorator
 
     def use_kwargs(self, *args, **kwargs):
         """Decorator that injects parsed arguments into a view function or method.
